% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_glmm_block_posterior.R
\name{fit_pois_glmm_block_posterior}
\alias{fit_pois_glmm_block_posterior}
\title{Poisson GLMM with block posterior}
\usage{
fit_pois_glmm_block_posterior(num_iter, elbo_tol = 1e-04, ...)
}
\arguments{
\item{num_iter}{number of iterations}

\item{...}{other parameters}
}
\value{
a list with fit objects
}
\description{
Poisson GLMM with block posterior
}
