% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data_pois_glmm}
\alias{get_data_pois_glmm}
\title{Get Data for Poisson GLMM}
\usage{
get_data_pois_glmm(...)
}
\arguments{
\item{...}{parameters to be passed to lme4}
}
\value{
list
}
\description{
Get Data for Poisson GLMM
}
