% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbfgs_pois_glmm_optim.R
\name{pois_glmm_optim_lbfgs}
\alias{pois_glmm_optim_lbfgs}
\title{Optimize Poisson GLMM with L-BFGS}
\usage{
pois_glmm_optim_lbfgs(par, par_scaling, glmm_data)
}
\arguments{
\item{par}{initial parameters.}

\item{par_scaling}{scaling for parameters to ensure proper conditioning}

\item{glmm_data}{data for GLMM problem}
}
\value{
optimized parameters
}
\description{
Optimize Poisson GLMM with L-BFGS
}
